Reorg: project reorganized
